fixed BUG with incorrect init value for info.least_rx_seq[i]

+++ vtun.h
@@ -464,7 +464,7 @@ struct phisical_status {
     /** Events */
     int just_started_recv; /**< 0 - when @see FRAME_JUST_STARTED hasn't received yet and 1 - already */
     int check_shm; /**< 1 - need to check some shm values */
-    unsigned long least_rx_seq[MAX_TCP_LOGICAL_CHANNELS]; // local store of least received seq_num across all phy
+    uint32_t least_rx_seq[MAX_TCP_LOGICAL_CHANNELS]; // local store of least received seq_num across all phy
 };
 
 struct conn_info {
